refactors code to separate out the prompt card
