+++ client/components/index.js
@@ -30,4 +30,5 @@ export {default as ExampleCheck} from './ExampleCheck'
 export {default as Page404} from './Page404'
 export {default as PostLesson} from './PostLesson'
 export {Instructions} from './Instructions'
+export {default as Read} from './Read'
 export {default as Optimization} from './Optimization'
